% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmri.tracking.R
\name{ArrayIndex}
\alias{ArrayIndex}
\title{wrapper function of ArrayIndex_C}
\usage{
ArrayIndex(dims, xrange, yrange, zrange)
}
\arguments{
\item{dims}{dims}

\item{xrange}{xrange}

\item{yrange}{yrange}

\item{zrange}{zrange}
}
\description{
wrapper function of ArrayIndex_C
}
