% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmri.tracking.R
\name{dist.line}
\alias{dist.line}
\title{get the distance between x and line(y,z)}
\usage{
dist.line(x, y, z)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{z}{z}
}
\description{
get the distance between x and line(y,z)
}
