% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmri.tracking-package.R
\docType{package}
\name{dmri.tracking-package}
\alias{dmri.tracking}
\alias{dmri.tracking-package}
\title{dmri.tracking: DiST - Diffusion Direction Smoothing and Tracking}
\description{
It provides functions to apply the deterministic tracking algorithm - DiST (Wong et al 2016) <doi:10.1214/15-AOAS880> and to plot tractography results.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vic-dragon/dmri.tracking}
  \item Report bugs at \url{https://github.com/vic-dragon/dmri.tracking/issues}
}

}
\author{
\strong{Maintainer}: Seungyong Hwang \email{syhwang@ucdavis.edu}

Authors:
\itemize{
  \item Raymond Wong \email{raywong@tamu.edu}
  \item Seungyong Hwang \email{syhwang@ucdavis.edu}
}

}
\keyword{internal}
