% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmri.tracking.R
\name{fiber.in.out}
\alias{fiber.in.out}
\title{assuming inc, outc are coordinates with the center of the voxel being (0,0,0)
compute the distance of the current fiber directon to each face of the current voxel
create a box with x,y,z dimens and divide by travel speed(direct)}
\usage{
fiber.in.out(inc, direct, dimens)
}
\arguments{
\item{inc}{inc}

\item{direct}{direct}

\item{dimens}{dimens}
}
\description{
assuming inc, outc are coordinates with the center of the voxel being (0,0,0)
compute the distance of the current fiber directon to each face of the current voxel
create a box with x,y,z dimens and divide by travel speed(direct)
}
