% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmri.tracking.R
\name{fiber.track}
\alias{fiber.track}
\title{fiber tracking function}
\usage{
fiber.track(
  iind,
  eig,
  loc,
  map,
  rmap,
  n.fiber,
  xgrid.sp,
  ygrid.sp,
  zgrid.sp,
  braingrid,
  max.line = 1000,
  nproj = 1,
  thres.ang = 0.5235988,
  vorient = c(1, 1, 1)
)
}
\arguments{
\item{iind}{iind}

\item{eig}{eig}

\item{loc}{loc}

\item{map}{map}

\item{rmap}{rmap}

\item{n.fiber}{n.fiber}

\item{xgrid.sp}{xgrid.sp}

\item{ygrid.sp}{ygrid.sp}

\item{zgrid.sp}{zgrid.sp}

\item{braingrid}{braingrid}

\item{max.line}{max.line}

\item{nproj}{nprog}

\item{thres.ang}{thres.ang}

\item{vorient}{vorient}
}
\description{
fiber tracking function
}
