% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmri.tracking.R
\name{get.out.vox}
\alias{get.out.vox}
\title{convert current voxel vector index to the 3d index in ROI}
\usage{
get.out.vox(index, cvox, braindim, vorient)
}
\arguments{
\item{index}{index}

\item{cvox}{cvox}

\item{braindim}{braindim}

\item{vorient}{vorient}
}
\description{
convert current voxel vector index to the 3d index in ROI
}
