% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmri.tracking.R
\name{proceed}
\alias{proceed}
\title{proceed}
\usage{
proceed(vox0, dir0, eig, rmap, n.fiber, thres.ang = 0.5235988)
}
\arguments{
\item{vox0}{vox0}

\item{dir0}{dir0}

\item{eig}{eig}

\item{rmap}{rmap}

\item{n.fiber}{number of fiber}

\item{thres.ang}{threshold of angular}
}
\description{
proceed
}
