% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmri.tracking.R
\name{project.proceed}
\alias{project.proceed}
\title{project proceed}
\usage{
project.proceed(
  inc0,
  vox0,
  dir0,
  loc,
  eig,
  rmap,
  n.fiber,
  braindim,
  dimens,
  nproj = 2,
  thres.ang = 0.5235988,
  vorient = c(1, 1, 1)
)
}
\arguments{
\item{inc0}{inc0}

\item{vox0}{vox0}

\item{dir0}{dir0}

\item{loc}{loc}

\item{eig}{eig}

\item{rmap}{rmap}

\item{n.fiber}{number fiber}

\item{braindim}{braindim}

\item{dimens}{dimens}

\item{nproj}{nproj}

\item{thres.ang}{thres angular}

\item{vorient}{vorient}
}
\description{
project proceed
}
