% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmri.tracking.R
\name{tractography}
\alias{tractography}
\title{Tractography}
\usage{
tractography(loc, vec, x_lim = NULL, y_lim = NULL, z_lim = NULL)
}
\arguments{
\item{loc}{loc}

\item{vec}{vec}

\item{x_lim}{xlim}

\item{y_lim}{ylim}

\item{z_lim}{zlim}
}
\description{
Tractography
}
