% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmri.tracking.R
\name{update.v.obj}
\alias{update.v.obj}
\title{Update number of fiber directions after v.smooth.bw}
\usage{
\method{update}{v.obj}(pre, res)
}
\arguments{
\item{pre}{an object from v.est}

\item{res}{an object from v.smooth.bw or v.smooth or any list(vec, update.ind)}
}
\value{
updated object for fiber directions
}
\description{
Update number of fiber directions after v.smooth.bw
}
