% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmri.tracking-package.R
\docType{package}
\name{dmri.tracking-package}
\alias{dmri.tracking}
\alias{dmri.tracking-package}
\title{dmri.tracking: DiST: Diffusion Direction Smoothing and Tracking}
\description{
This package provides functions to apply the deterministic tracking algorithm and to plotting tractography result.
}
\author{
\strong{Maintainer}: Seungyong Hwang \email{syhwang@ucdavis.edu}

Authors:
\itemize{
  \item Raymond Wong \email{raywong@tamu.edu}
}

}
\keyword{internal}
