% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmri.tracking.R
\name{tractography}
\alias{tractography}
\title{Tractography}
\usage{
tractography(loc, vec)
}
\arguments{
\item{loc}{Voxel coordinates that the selected reconstructed fiber went through}

\item{vec}{Diffusion direction that used to reconstruct a fiber}
}
\description{
Visualize the result from v.track into 3D object.
rgl package is required.
}
\examples{
#Load example output from peak detection algorithm
#load(system.file("extdata", "peakresult.rda", package = "dmri.tracking"))

#str(peak.result)  #Output from the peak detection algorithm

#Apply Tracking algorithm
#result = v.track(v.obj = peak.result, max.line=500)


#To plot the tractography based on the previous tracking result,
#comment out the following lines and run them.

#library(rgl)
#open3d()
#for (iind in (result$sorted.iinds[result$sorted.update.ind])){
#  cat(iind,"\n")
#  tractography(result$tracks1[[iind]]$inloc, result$tracks1[[iind]]$dir)
#  tractography(result$tracks2[[iind]]$inloc, result$tracks2[[iind]]$dir)
#}

#More detailed procedure is available in (https://github.com/vic-dragon/dmri.tracking)

}
\seealso{
The tracking result from \code{\link{v.track}} can be used for \code{\link{tractography}}
}
