% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmri.tracking.R
\name{tractography}
\alias{tractography}
\title{Tractography}
\usage{
tractography(loc, vec)
}
\arguments{
\item{loc}{The list of voxel coordinates that the selected reconstructed fiber went through}

\item{vec}{The list of diffusion direction that used to reconstruct a fiber}
}
\description{
Visualize the result from v.track into 3D object.
rgl package is required.
}
\examples{
#Load example output from peak detection algorithm
#load(system.file("extdata", "peakresult.rda", package = "dmri.tracking"))

#peak.result  #Output from the peak detection algorithm

#Apply Tracking algorithm
#result = v.track(v.obj = peak.result, max.line=500)

#library(rgl)
#open3d()
#for (iind in (result$sorted.iinds[result$sorted.update.ind])){
#  cat(iind,"\n")
#  tractography(result$tracks1[[iind]]$inloc, result$tracks1[[iind]]$dir)
#  tractography(result$tracks2[[iind]]$inloc, result$tracks2[[iind]]$dir)
##}
}
