% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmri.tracking.R
\name{v.track}
\alias{v.track}
\title{Deterministic tracking algorithm from DiST by Raymond et al.(2016)}
\usage{
v.track(
  v.obj,
  max.line = 100,
  nproj = 1,
  elim = T,
  elim.thres = 1,
  thres.ang = 0.5235988
)
}
\arguments{
\item{v.obj}{The list type of object from the peak detection algorithm. The object should contain the following components.
\itemize{
 \item{vec:} Estimated peak direction.
 \item{loc:} Coordinates in braingrid for each estimated peak direction.
 \item{map:} Indicator of voxel for each estimated peak direction
 \item{rmap:} Starting location in map of each voxel
 \item{n.fiber:} Number of detected peaks at each voxel
 \item{n.fiber2:} Number of detected peaks corresponding to map
 \item{braingrid:} Normalized voxel coordinates based on the number of voxel and the size of each axis.
 \item{xgrid.sp,ygrid.sp,zgrid.sp: } Voxel size of x, y, z-axis, respectively. (e.g. Voxel size in HCP dMRI: 1.25mm * 1.25mm * 1.25mm)
}}

\item{max.line}{Maximum number of voxels that the reconstructed fibers can go through}

\item{nproj}{Considered number of neighborhood voxels if the algorithm cannot find no viable direction nearby.}

\item{elim}{Choice of whether the reconstructed fiber is removed from the result if the length is less than elim.thres.}

\item{elim.thres}{If elim=TRUE, a lower limit of the reconstructed fiber length.}

\item{thres.ang}{Threshold to determine whether the destination voxel have a viable direction. (default value: pi/6). That is,
the algorithm is proceeded, if the angular difference of the diffusion direction between the previous voxel and the destination voxel is smaller than thres.ang}
}
\value{
Result of deterministic tracking algorithm
\itemize{
 \item{v.obj:} Input of tracking algorithm.
 \item{track1, track2:} reconstructed fiber information
 \itemize{
  \item{inloc:} The voxel coordinates that the reconstructed fiber went through
  \item{dir:} The diffusion direction that used to reconstruct fiber
  \item{iinds:} The indicator of voxel that the reconstructed fiber went through
  \item{change:} Indicator that which voxel could not pass the angular difference threshold (thres.ang)
  \item{pvox:} The passed voxel
  \item{pdir:} The passed direction
  \item{pdis:} The length of fiber between the passed information and the previous one.
  \item{ppdis:} The distance of the voxel between the passed one and the previous one.
 }
 \item{n.iinds:} Number of voxels the reconstructed fiber went through
 \item{n.use.iind:} Number of times that each estimated direction is used.
 \item{lens:} The reconstructed fiber lengths.
 \item{sorted.iinds:} The ordered reconstructed fibers
 \item{sorted.update.ind:} Whether the reconstructed fiber is greater than elim.thres
}
}
\description{
v.track applies the deterministic tracking algorithm, DiST, to reconstruct neuronal fibers based on the peak detection results.
}
\examples{
load(system.file("extdata", "peakresult.rda", package = "dmri.tracking"))
peak.result  #Output from the peak detection algorithm

result = v.track(v.obj = peak.result, max.line=500, nproj=1, elim=T, elim.thres=1, thres.ang=0.5235988)
}
